Extracted base tooltip Removed fallbackPosition prop
