+++ packages/components/src/base_tooltip/constants.jsx
@@ -5,3 +5,9 @@ export const POSITION_RIGHT = 'right';
 
 export const TOOLTIP_OFFSET = 10;
 export const TOOLTIP_FLIP_SCREEN_OFFSET = 10;
+
+export const TRIGGER_HOVER = 'hover';
+export const TRIGGER_CLICK = 'click';
+export const TRIGGER_DELAYED = 'delayed';
+
+export const DELAY_TIMEOUT = 1000 * 3;
