--- packages/components/src/feature_alert/constants.jsx
@@ -1,5 +1,0 @@
-export const TRIGGER_HOVER = 'hover';
-export const TRIGGER_CLICK = 'click';
-export const TRIGGER_DELAYED = 'delayed';
-
-export const DELAY_TIMEOUT = 1000 * 3;
