+++ packages/components/src/feature_alert/index.stories.jsx
@@ -1,18 +1,16 @@
 import React from 'react';
 import { text, select, boolean, withKnobs } from '@storybook/addon-knobs';
 import { action } from '@storybook/addon-actions';
-import {
-  Tooltip,
-  TRIGGER_HOVER,
-  TRIGGER_CLICK,
-  TRIGGER_DELAYED,
-} from './index';
+import Tooltip from './index';
 import {
   POSITION_TOP,
   POSITION_BOTTOM,
   POSITION_LEFT,
   POSITION_RIGHT,
-} from '../tooltip/constants';
+  TRIGGER_HOVER,
+  TRIGGER_CLICK,
+  TRIGGER_DELAYED,
+} from '../base_tooltip/constants';
 import styles from './index.stories.module.scss';
 
 export default { title: 'Feature Alert', component: Tooltip, decorators: [withKnobs] };
@@ -37,7 +35,6 @@ export const Default = () => (
     <Tooltip
       content={text('Tooltip text', TOOLTIP_PLACEHOLDER)}
       position={select('Position', POSITIONING_OPTIONS, POSITION_TOP)}
-      fallbackPosition={select('Fallback position', POSITIONING_OPTIONS, POSITION_TOP)}
       trigger={select('Trigger', TRIGGER_OPTIONS, TRIGGER_HOVER)}
       closeIcon={boolean('Has close icon', true)}
       defaultOpen={boolean('Is default open', false)}
