+++ packages/components/src/tooltip/index.stories.jsx
@@ -7,7 +7,7 @@ import {
   POSITION_BOTTOM,
   POSITION_LEFT,
   POSITION_RIGHT,
-} from './constants';
+} from '../base_tooltip/constants';
 
 const TOOLTIP_PLACEHOLDER = 'Design is like a joke.';
 
@@ -24,7 +24,7 @@ const POSITIONING_OPTIONS = {
 export const Default = () => (
   <div className={styles.container}>
     <Tooltip
-      type={select('Tooltip position', POSITIONING_OPTIONS, POSITION_TOP)}
+      position={select('Tooltip position', POSITIONING_OPTIONS, POSITION_TOP)}
       bubble={text('Tooltip text', TOOLTIP_PLACEHOLDER)}
     >
       {`Tooltip position: ${select('Tooltip position', POSITIONING_OPTIONS, POSITION_TOP)}`}
